# openapi/model-service-v2.yml
openapi: '3.0.0'
info:
  title: 'Model Service'
  description: 'Model Service v2 API - This API is under active development and subject to change.'
  version: '0.0.1'
servers:
  - url: https://api.pennsieve.io/models/v2
  - url: https://api.pennsieve.net/models/v2
paths:
...
  '/organizations/{organizationId}/search/records':
    post:
      tags:
        - Search
      summary: 'Search records across datasets'
      operationId: server.api.v2.search.records
      parameters:
        - name: X-Bf-Trace-Id
          in: header
          schema:
            type: string
          required: true
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: limit
          in: query
          description: limit
          required: false
          schema:
            type: integer
            format: int64
            default: 25
        - name: offset
          in: query
          description: offset
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: orderBy
          in: query
          description: Order by property
          required: false
          schema:
            type: string
            maxLength: 32
            nullable: true
        - name: orderDirection
          in: query
          description: Order by direction
          required: false
          schema:
            type: string
            enum: ["asc", "desc"]
            default: "asc"
      requestBody:
        description: The search request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordSearchResult'
      security:
        - Bearer: []
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: server.auth.decode_token
  schemas:
    OrganizationIdentifier:
      type: integer
...
    SimpleDataType:
      type: string
    ComplexDataType:
      type: object
    DataType:
      oneOf:
        - $ref: '#/components/schemas/SimpleDataType'
        - $ref: '#/components/schemas/ComplexDataType'
    GraphValue:
      anyOf:
        - type: boolean
          nullable: true
        - type: array
          items:
            type: boolean
          nullable: true
        - type: string
          nullable: true
        - type: array
          items:
            type: string
          nullable: true
        - type: number
          nullable: true
        - type: array
          items:
            type: number
          nullable: true
...
    Property:
      type: object
      required:
        - name
        - displayName
        - dataType
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        dataType:
          $ref: '#/components/schemas/DataType'
        index:
          type: integer
          format: int64
        locked:
          type: boolean
        default:
          type: boolean
        modelTitle:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        description:
          type: string
        required:
          type: boolean
          default: false
...
    RecordStub:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
    Record:
      type: object
      required:
        - id
        - values
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
        values:
          type: object
          additionalProperties:
            anyOf:
              - $ref: '#/components/schemas/GraphValue'
              - $ref: '#/components/schemas/RecordStub'
                nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
...
    SearchRequest:
      type: object
      required:
        - model
        - filters
      properties:
        model:
          type: string
        datasets:
          type: array
          items:
            type: integer
            format: int32
        filters:
          type: array
          items:
            $ref: '#/components/schemas/PropertyFilter'
    PropertyFilter:
      type: object
      required:
        - model
        - property
        - operator
        - value
      properties:
        model:
          type: string
        property:
          type: string
        value:
          oneOf:
            - type: boolean
            - type: string
            - type: number
        operator:
          type: string
          enum: ["=", "<>", "<", "<=", ">", ">=", "STARTS WITH", "CONTAINS", "IS"]
        unit:
          type: string
    RecordSearchResult:
      type: object
      required:
        - models
        - records
        - totalCount
        - limit
        - offset
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultModelStub'
        records:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Record'
              - type: object
                required:
                  - modelId
                properties:
                  modelId:
                    type: string
                    format: uuid
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
    SearchResultModelStub:
      type: object
      required:
        - id
        - properties
        - dataset
      properties:
        id:
          type: string
          format: uuid
        dataset:
          $ref: '#/components/schemas/Dataset'
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    Dataset:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
        nodeId:
          type: string
          nullable: true
...